solve the itemfix bug
using dtlTraders will, with several usage cases, restore the health of all the
items in the player's inventory. this commit attempts to solve this bug
in one major usage case: when a regular player interacts with the trader
NPC.

some usage cases remain but they all occur in management mode, so
the impact is way smaller.